move TestCalculateStakeModifier to the consensus package where it belongs
+fix ineffassign in the consensus module
+add ./modules/blockcreator to the testable packages
